Merged PR 16851: Refactored print log + counting function to suit both french receipts and every other receipt
This PR reverts a timing change of the print log that was done in a hurry to comply with french compliance, which ended up breaking the print counter for all other prints outside french.

It also adds a new auxiliary function that the french receipt can use instead to get a count that suits the compliance without impacting other receipts.

The french receipt template has been updated to fit the changes in this PR.

Related work items: #10078