Merged PR 16915: Printnode enum field default value
Customers that had an older printnode config in their BC (from v9 or older) was getting invalid field values 0 in the two new fields after being updated.

I was not aware enums are not initialized to first or lowest integer, seems like a BC bug it isn't but we need a fix regardless so i added an upgrade codeunit that guarantees it.

Also locked two labels in the report printer subscribers we have, to prevent further complications from those.

This was tested by setting up printnode on v9 and deploying this branch with launch.json "forceUpgrade: true". The printnode records now look good and "printer selections" dropdown works as well.

Related work items: #10234