Merged PR 16817: Fixed issue with primary ticket request not creating a ticket.
Only when a ticket request has "Admission Inclusion" as "Required", a ticket will be created and also only for the primary request line.

With dynamic tickets this was a little out of sync - the primary request line was assigned to an optional admission - so no ticket was created.

Related work items: #10052