Merged PR 16906: Case 564157 - Overflow warnings
Case 564157 - Overflow warnings

**NOTE:** File StoresWP.Page.al, procedure CheckIfNoAvailableInMagStore
It would create more issues with overflow if I just changed the length of parameter.
I had to create new local variable to use in the other functions that are length 10 for this parameter, and I didn't want to put the hardcoded length 10 for the first issue. **Logic is not changed.**

Please contact for more details/discussion.

Related work items: #10228