Merged PR 17123: 565980 - HTML Display validation fix + Html list selection
This fix includes the fix for HTML Display, the input validation page did not show when ending a sale and going to Login page. Since the code used Current sale to identify POS Unit. Now it uses Session.

There is added an action to the POS HTML Display profile that functions almost the same as POS Menus Download Template Data. A List of HTML on our Blob storage is displayed and the user can select a standard html from this list.

Related work items: #10916