Merged PR 17254: Improve Member Services performance with conditional execution of SelectLatestVersion()
Currently every request to member services clears the NST cache (SelectLatestVersion) since there is a problem when 2 consecutive messages are routed to different NST's.

Depending on the time between the messages, NST 2 does not immediately see changes committed from NST 1. The "cure" is to SelectLatestVersion() which invalidates the NST's cached data and forces it fetch again. This naturally comes with a cost.

The solution here is to include the NST instance ID in the response data so client can included it in the next request sent to server.

When server sees the same id in the request as is return by the ServiceInstanceId() function, server can skip executing the SelectLatestVersion() since data in request is based on data generated by this instance.

Empirical tests suggested that 30 memberships (4 messages each) dropped from 19-ish seconds to13-ish seconds in total execution time when SelectLatestVersion was not executed (hardcoded) in an container dev environment prior to this PR. This was repeatable.

Testing with PR the execution times goes all over the place and up to now have not been stable enough to make a conclusive decision if it is worth it.

Related work items: #11035