Merged PR 17391: Remove Magento 2 Payment Gateway module
In [this task in the Magento 2 project](https://ac.navipartner.dk/projects/magento-2-1/tasks/257) (case 355841) we implemented some custom functionality where BC could call Magento and have the web shop capture the payment.

After discussing with the backend developers, we can see that this module is not included in any of the NP Ecommerce web shops anyway, so to make sure we don't have to keep it up to date going forward.

Case references: 555854, 568373

Related work items: #11120