Merged PR 17502: Optimizing Magento Websites Page layout part 1
Optimizing Magento Websites Page layout part 1 (Part 1 is because page is public, so we need first to put grid to obsolete pending and then to remove it in 2nd step).

After it's removed, page will look like this:

![Variant 2.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/17502/attachments/Variant%202.png)

Group is added to separate stores from websites in a nice way. There is a horizontal line between.

Related work items: #11115