Merged PR 17717: WSDL pretty print
This is a bit of an experiment.

I got hold of a xml transformation script to render a WSDL into a human readable format. WSDL == SOAP API documentation intended for a computer.

But I always get the API documentation question so I would like to try this out.

(See the copyright message in wsdl-viewer.xsl)
I could re-factor this file into something more NP-ish, but ... do I have to?

Related work items: #11355