Merged PR 17742: Creating APIs, add events and refactor code in order to be compactible with Group88
Added new events with the handled pattern.
Expose some of the procedures through API and refactor code regarding Item Worksheets and Sales Price
Maintenance.

Just to put context into this PR. NPCore was upgraded from version 9.x to 18.x on the Group88 project without checking if their ZAL extension was compatible with version 18.x. And the other problem is that Group88 is still on version 17 of Business Central and they are using the old pricing module.

Related work items: #11363