Merged PR 17805: Moved from pseudo obsoleted option to actually obsoleted enum
Moved from pseudo obsoleted option to actually obsoleted enum.
The main problem is that an old field in the primary key of POS Sale Lines is obsoleted.
But you cannot mark it as obsolete to give everyone warnings when it's of type option. By using enums instead we at least get warnings.

This PR is mainly an experiment to see if the code cops will allow it. If yes, it should make the solution a tiny bit more intuitive.