Merged PR 17516: Create interface for Payment Gateway integrations
This PR introduces an interface to replace the old event subscriber based module.

Now it is more clear to integrations what they should, and in the long run, we will be able to make `NPR Magento Payment Line` and `NPR Magento Payment Gateway` internal (or even moved out of the relation to Magento) so we get a really crisp API towards partners and PTEs.

The PR is very large so I'm looking for feedback in the following areas:
- Upgrade Code
- Central functionality like capturing, refunding, and cancelling, i.e. the changes made to `NPR Magento Pmt. Mgt.` and relevant Try-Catch codeunits.

Related work items: #8880