Merged PR 18040: UTC time not correct ticket XML messages
The start and finish time for ticket validity was communicated incorrectly to web using UTC time without also providing the time zone.

Rather than going down the time zone hole and to keep it simple, 4 new fields are added
* start_date
* start_time
* end_date
* end_time

these dates and times are the local time. There is no time zone provided.

3 XmlPorts are affected. SOAP Actions are:
* ConfirmTicketReservation
* ListTickets
* ConfirmTicketChangeRequest

This is made such as it is not a breaking change. The original fields are still included in the XML but the implicit time zone is UTC 0.

Related work items: #11586