Merged PR 18068: Printnode duplicate id in report printer subscriber bug
Added hashset to avoid adding id multiple times in the subscriber.

This works fine because the call to printnode will check if the report ID has any specific record and if yes, it will use those. So it already has correct filtering, so no need for us to add the same inside "Printer Selection".

This also means we don't have to upgrade any data.

Related work items: #11607