Merged PR 18234: Arithmetic overflow when posting many entries
Reset line number to 0 on each posting run.

Line Number is an integer and is increated with 10000 with each journal line being created.
2147483647 (2^31-1) / 10000 = 214747

LineNumber is a global in the posting codeunit and when executed many times in a row by JQ it will overflow as it will not reset to zero between runs.

Thus when there are more than 214747 journalines created (across the JQ workspace for 1 session) the overflow would occur.

Related work items: #11717