Merged PR 18300: Skip callbck to case sys. if Evaluat, demo or test
Skip callback to case system if Evaluation, demo or test.

I have added next 3 rules:

1. To skip if Company is set as "Evaluation Company" (this is standard field of course and it's automatically enabled for "CRONUS" companies for example).
2. To skip if record in our "NPR Environment Information" table doesn't exist. I have added this because it could happen that on first login to some tenant code for callback could be called before code for inserting record in "NPR Environment Information". (Because it is same Event).
3. To skip if "Environment Type" field in table "NPR Environment Information" is set to DEMO or Sandbox/Test/Development

By these three we will get rid of some number of callbacks but not for all of non production ones. As I already wrote two times, we still have many companies which are not production ones, but the are marked as PROD in our "NPR Environment Information" table and NOT marked as "Evaluation Company" in Company table. These are examples and I doubt that any of these companies is really production company:

![Non production 1.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/18300/attachments/Non%20production%201.png)
![Non production 2.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/18300/attachments/Non%20production%202.png)

I could go as I already explained earlier and do check if company name contain 'test', 'live', 'old', 'restore', 'setup', etc... but this is NOT safe approach. I am not saying that they would or will, but customers could misuse that and name their companies like that and skip paying for it.

OK, we could find some idea to doublecheck on that, like Mikkel suggested to count Max active POS sales, max POS sales per day, max POS posting entries per day etc. And if amount is higher or some value, we will still consider it as production.

But what if user create a copy of production company (in order to test something or just to backup it) and name it as test, or take empty one and do rapid start import to fill the data to do some test. That company would still be treated like production, although it's test.

Related work items: #11748