Merged PR 18594: Add Actionable attribute to Nc Import Types
Over the years it has become apparent that our use of logging in the import list is confusing both consultants and customers.

This PR introduces an attribute that determines if import entries of a given type is 'actionable' meaning that a user can take action after the fact to resolve the import issue. The default is `true` to keep the original behavior.

Related work items: #11974