Merged PR 18686: Import list, display error message in a window
Import list, display error message in a window.

Instead of empty window:

![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/18686/attachments/image.png)

Now we will have same message in the window as in Error Message.

P.S. When it comes to Account authentification like in this particular example, in next PR I will change that error handling to be more detailed than this "Contact not found"

Related work items: #12012