Merged PR 18359: Fix for getting the right MixDiscount
Fix for getting the right MixDiscount

This is a fix where for cases when wrong MixDiscount is taken for POS Sales.

This happens in scenarios where Mix Discounts have more than 1 Item on the lines, where the system (because of a bug) would use wrong Item for filtering out the best Mix Discounts to be used.
This is caused because Mix Discount Lines are only filtered by Mix Discount and not by Item No. for which they are used so mix-ups create bad matches.

Fix for the problem is applied in codeunit 6014416 “NPR Mixed Discount Management” and procedure MatchMixedDiscount where additional filter for the MixDiscountLine is added to always filter the Item for which the discount is applied.
This fixes the real problem where CALCSUM of Quantity on function FindBestMixMatch is returning 0 because the POSSaleLine is filtered by wrong Item so no matches can be found.

Related work items: #11420