Merged PR 18783: Contact No. - String Lenght Exceeded
In table "NPR POS Sale" field "Contact No." is text [30] . It is used to store Contact Name (as Attention in Sales_Doc_Exp) in Sales Document
![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/18783/attachments/image.png)

and after leaving out Customer Type it is used to store Contact.No with Contact_Select POS Action:
![image (2).png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/18783/attachments/image%20%282%29.png) .

FieldName is misleading and caused runtime Error with Contact.Get('value text over 20').
This commit is a fix but this looks like something that needs to be refactored.

![StringLenghtExceeded.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/18783/attachments/StringLenghtExceeded.png)

Related work items: #12081