Merged PR 18851: TinyMCE Text Editor use entire ControlAddIn area
TinyMCE Text Editor use entire ControlAddIn area.

**Before** it was like this:

![before copy (2).jpg](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/18851/attachments/before%20copy%20%282%29.jpg)

**With the change in this PR is:**

![playing with 7 copy (2).jpg](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/18851/attachments/playing%20with%207%20copy%20%282%29.jpg)

The difference is not huge, but it's better, because user doesn't have to catch slider in bottom right corner and pull it down.

Except this change, I have tried many ways to make this real full screen like we have POS for example (and I managed), but in that case we don't have close button. And I doubt that we have possibility to change this Addon by adding our code to it. (maybe we have, I haven't tried. I didn't wanted to spend more time).

Different option than to full screen, was to have ControlAddin in complete area (but keeping BC upper toolbar) and of course close button. It was like this:

![Try fullscreen 2 (2).png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/18851/attachments/Try%20fullscreen%202%20%282%29.png)

But... "form" element cannot be changed from AL code. Neither CSS style can be added to it. I have found some blog where someone asked for the same thing, but reply is that it's not allowed/supported.

At the end...
@<Mark Stewart Pedersen> if this is not enough still, in a couple of minutes I could add one more change (I have tried it already) and it would look like this:

![could be like this 2 (2).png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/18851/attachments/could%20be%20like%20this%202%20%282%29.png)

but personally, I don't like this option. We do get a bit more workspace, but it pops out of the ControlAddin frame

Related work items: #12130