Merged PR 18992: fixed order of conditional and filter on Document No instead of POS Unit, thi...
The OnAfterEndSale is triggered and ran a filter and getlast query on the POSEntry based on the POS Unit, now it does the same but on Document No. for better performance.

Also the order of condiftionals is shifted so it checks for the POS HTML Profile first, so it will exit faster.

Related work items: #12282