Merged PR 19239: Added TMTicketAdmissionData public CU
Added TMTicketAdmissionData public codeunit, based on case 578287

I haven't added 'Access = Public' property because default value is Public. (I don't know what is 'the rule', are we adding it always or not)

If you wish to name it differently, or to change naming of JSON keys to 'admission_code' instead of 'admissionCode' etc, I will change.

Or anything else... ;)

Related work items: #12269