Merged PR 19328: Apply discount on coupons - ITEM_LIST
Additional development: when the coupon is applied on sale and maximum qty is filled, new field "Apply discount on items with" is created for options highest price, lowest price and priority of item on the list.

Bug fixed: discount changes when the coupon is applied and qty of item is changed.

Related work items: #12097