PR 19325: Use Record ID and Record Position to link tables in print templates
There was no way to link a related table in print templates, if the linking needed to be based on record ID or record position of linked or parent table.
Hereby introducing new fields `Parent Link On` and `Link On` (enums with values "Field", "Record ID" and "Position") to identify based on what links are to be build.
We can now use record id or position of the record both for parent and linked tables.
The functionality is available for defining both data item links and data item constraints.

Case 567862
Related work items: #12461