Merged PR 19350: Assign Serial No to BOM components on the POS
New POS Action Created - ASSIGN_SERIAL_NO it can be used to assign a serial no to the current line on the POS. The new action has a parameter - Select Serial No - if it's set to true and the item is setup to have a full serial no tracking, then user will have to select a serial no from a list.
The POS Action Item - has been modified to take into account the cases when a BOM Item is exploded on the POS and the component items require a serial number. In that case the system is going to ask the user to enter a serial number for each component. The user can cancel the serial no assignment and add the serial numbers at a later stage with the new action.

Related work items: #12479