Merged PR 19414: Changes to ListTicketItem service
Added additional information to to service:

In item section the following subsection was added:
<code><ticket_description></code>

containing the following elements:
<code>
      <title/>
      <subtitle/>
       <name/>
       <description/>
       <full_description/>
</code>

Field values are selected according to configuration on ticket setup - shared with how data is communicated with ticket server.

The same subsection was also added under element <code><admission_description><code> for an admission that is optional and has a configured item as it source for price.

<code>item_number</code> attribute was added to <code>admission_is</code> element.

Related work items: #12538