Merged PR 19439: Case: 583141 Refresh SALES_DOC_EXP and SALES_DOC_IMP Menu Action Parameters o...
Case: 583141 Refresh SALES_DOC_EXP and SALES_DOC_IMP Menu Action Parameters on Upgrade

We have added new parameters to the following actions

SALES_DOC_EXP
 -GroupCode
 -GroupCodeEnabled

SALES_DOC_IMP
 -GroupCodeFilter
 -GroupCodeFilterEnabled

When creating a new container for version BC22 the MenuButtonParameters were not updated for those actions and the users had to refresh the actions manually.
The problem doesn't occur in other BC versions.

![image.png](https://dev.azure.com/navipartner/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/19439/attachments/image.png =750x)

Because of that I have created an upgrade procedure that refreshes the parameters of the of the Menu Buttons that are related to those actions.

Related work items: #12556