Merged PR 19443: End of Day concurrency issue
This change addresses a concurrency issue in end of day process.

Main issue was a deadlock caused when iterating to insert pos entry lines for balancing in <code>CreateBalancingEntryAndLines()</code> The <code>FindSet()</code> caused a nasty SQL statement when doing <code>Next()</code> in the loop. It was replaces with a <code>Query</code> to make the loop itself non-blocking.

A <code>Commit()</code> was moved to be generic rather than specific for UI mode.

Tests added to run EOD concurrence via SOAP as it is easier to get telemetry and debug options.

This required a small change to the constructor of POSSession to also initiate POS Unit. The statement just before initiates Setup with POS Unit, but it was not carried into POS Session, this must have been a mistake.

Related work items: #12558