Merged PR 19411: Refactor process of applying discounts, remove unnecessary code anf fix old bugs
The process of applying coupons with the highest and lowest prices was not working so in this PR it is refactored process of using those coupons, introduced with temp tables.
There was also a dead code which was removed, also removed unnecessary code, and fixed bugs when changing the quantity of items with applied coupons. Earlier the coupon % was not changing as it should.

Related work items: #12097