Merged PR 19542: Uable to complete a 0 amount POS Sale
Case: 583986

Complete a 0 amount POS Sale after payment

If the POS Sale amount is 0 and you try to pay it the system is not completing the POS sale.
This a problem in the cases where you need to issue a 0 amount voucher or if you need to renew a membership etc.

Related work items: #12636