Merged PR 19453: Random order when printing exch labels
Case: 580423 Random order when printing exch labels

When printing exchange labels with option All from the POS the system was printing the labels in a random order.
The reason for that was that we were sending the print jobs asynchronously to the printer and sometimes thy were printed in the right order and sometimes they werent.
I have made changes to the printing process flow to address the issue.

Related work items: #12564