Merged PR 19699: Case: 576663 Cannot delete cash payments
Case: 576663 Cannot delete cash payments

Check POS View when deleting lines on the POS.

There was a bug where you were not able to delete the payment lines if there were no sales lines on the POS.
The problem was caused because we were not checking the current POS View and we were always checking if the sales line were valid.
Now if you're in the POS Payment View the system is going to delete the payment lines.

Related work items: #12731