Merged PR 19682: Fiscal Serbia: Journal Field Text to Media, Vouchers fiscal., Voucher tests, Automated tests Mock response
Hi Mikkel, everything is same as on pervious PR which I abandoned in order not to spam you with notifications.
So as we discussed there I created automated tests for voucher fiscal. behaviour
Also I refactored all RS tests not to use http request and to use Mock
Everything else is same as on pervious PR

Related work items: #11851, #11999, #12276, #12568, #12569, #12644