Merged PR 19739: Case: 584611 telem
Case: 584611 telem

When the balancing was done from the logon screen the system was searching
for an EFT Setup for the POS Unit it the POS Sale, but since we were not in
a POS Sale - the POS Unit was blank. Because of that the system was searching
for a EFT Setup with a blank POS Unit.

Related work items: #12752