Merged PR 19893: Fix error to modify existing line, and remove true parameter when modifyiling...
There was an error in the code where the Insert function is called instead of Modify.

The other part of this PR is removing the RunTrigger parameter set to true when modifying Price List Lines, it was changed because if the RunTrigger is applied, standard code for changing status to Draft will be applied. We do not want that type of behavior when changing already Active lines, the expected behavior is to status stays the same as it was.

These issues are originally found in case 585620 for the Group88 client.

Related work items: #12935