Merged PR 19620: Migrate voucher payment flow to PAYMENT_2
We have migrated the vouchers and the external vouchers to PAYMENT_2 module.

For internal vouchers codeunit 6151444 "NPR POS Action Scan Voucher2" now implements the "NPR IPOS Workflow" interface.
For foreign vouchers codeunit 6059939 "NPR POSAction ForeignVoucher" implements the "NPR IPOS Workflow" interface.
I have moved the HideAmount and HideZeroAmountDialog parameters from the payment_2 action to the POS Payment Method table, because it was not possible to use the parameters on a Menu button when you set up the button as a payment type.

On the POS Payment Method Card there is a new field called Hide Amount Dialog. It is an enum field with the following options:

Never - the system is never going to hide the amount dialog on the POS
Always - the system is going to always hide the amount dialog on the POS
Zero Amount - the system is going to hide the amount dialog on the POS when the amount is 0

![image.png](https://navipartner.visualstudio.com/4337df0c-8118-400a-8f4f-dc29e226a64a/_apis/git/repositories/7ca1eecf-6774-4272-9cba-1dfd3ba208e3/pullRequests/19620/attachments/image.png)

Related work items: #12676