Merged PR 18740: Send POS Stores and Unit qty. upd. to CS frm SaaS
Send POS Store and POS Unit quantity updates to case system when there is a change in number of POS Units or POS Stores in SaaS client environment.

Before we used Tenant Diagnostic as common table for SaaS and non SaaS environment.
And when we created new callback, dedicated to SaaS only, we (actually I) have created new set of tables for SaaS, but I totally forgot to change this part. I have noticed it while working on case 576997.

Related work items: #12049