Merged PR 20078: Fail on unkown item reference in NpEc import
Previously the code would just exit thereby importing the order in a faulty state. This PR changes this behavior such that is throws an error if it cannot translate a reference to an Item Variant.

This was already implemented in C/AL but was never moved to AL.

Related work items: #13015