Merged PR 20103: #12936 Sentry Cron Monitoring
**codeunit "NPR Sentry Cron"**
 - new codeunits which keeps the logic for using Sentry Cron's APIs for creating check-ins and updating check-ins
 - at the moment this should work only SaaS production environments

**codeunit "NPR POS Post Item Entries JQ"**
 - updated it to create Sentry Cron heartbeats

**codeunit "NPR POS Post GL Entries JQ"**
 - updated it to create check-in and then to update it with appropriate status

Note: Talked with Mikkel and we have decided that for the time being we will be creating monitors with hardcoded values. Here we have set Schedule, Check-in Margin in Minutes and Max Runtime in Minutes for Cron monitors based on our default values for the Job Queue Entries for related codeunits.

Related work items: #12936