Merged PR 20081: Reverse direct sales POS action - global dimensions on lines
Using reverse direct sale POS action, global dimensions on POS sales lines were always copied from the original sales ticket (the one that you want to reverse) even though the parameter "copy header dimensions" is false.
I added an additional parameter "copy line dimensions" and this part of the code will be executed only if the new parameter is true. The parameter will be set to true by default and it will work as it is now, so it won't affect the current POS setup on client dbs.
For the client Group88 the new parameter "copy lines dimensions" will be set to false and the old parameter "copy header dimensions" to false, so dimensions will be taken from the current POS unit and not from original pos sales line.

Related work items: #13003