PR 20220: Rest.module: not supported line handling, when saving POS sale to waiter pad
We cannot save to waiter pads just any lines added to POS sales. Namely, only lines of type "item" and "comment" can be saved to waiter pads.
However, system did not check that, and in those cases system did not know how to restore those kinds of lines back to POS sale. This was specifically dangerous with payment lines, which essentially could be just lost.
Hereby introducing a new functionality, which is going to check line compatibility and prevent saving and deletion of non-compatible lines.
3 POS actions have been affected:
1. "RUN_W/PAD_ACTION"
1. "SAVE_TO_WAITER_PAD"
1. "RV_SET_R-VIEW"

Case 587991
Related work items: #13190