Merged PR 20317: Conditionally suppress EOD posting for virtual bins
This setting only applies to virtual bins and prevents the aggregated result from being posted. This is useful when EFT bank reconciliation is done per entry.

It can only be allowed for virtual bins as there will not be a difference to handle.
Original posting (from EFT) is the truth.
Original posting is also on the final account
Actual reconciliation is done outside of the end-of-day process.

Related work items: #13298