Merged PR 19551: 533264 - SFTP Out of memory exception
This fix introduce the concept of file server profiles, which is a data table that describes a connection to a file server. This will reduce redundant data on file server info and make it easier to work with.
The SFTP Azure function ran out of memory when downloading files bigger than ~75MB and to fix this problem main structural changes needed to be made for the SFTP. Now it will be easier to add new versions to SFTP and deprecate old ones.

Also updated the 2 places where FTP/SFTP were used, and introduced the file server profile concept there. These changes will not break existing functionality as the old azure functions and codeunit are intact but will be marked as obsolete pending.
Fileserver profile will take presidence over the preexisting ftp server settings, if such profile is declared.

A new setup has also been made for Azure function apps, where new apps can be added to the new structure, and make it easier to see which functions are used. This setup should be expanded to all functions in the long run.

Related work items: #12641