Merged PR 20328: Web Service discount_coupon_service returning POSSession not initialized error
When the web service discount_coupon_service was called the system was
returning an error that the POSSession was not initialized. This was due to us adding a check after the web service was designed. The current pr initializes the POSSession in order to cater for the problem.

Related work items: #13311