Merged PR 20472: Optimization of POS Delete Action to call the server only one time.
Optimization of the DELETE_POS_LINE action which allows it to call the server only one time.
In order for this to happen the workflow.scope.view global variable has been used in the javascript file, which allows the view type to be checked without calling the server.

Related work items: #13417