Merged PR 20497: Ask for Amount in Partial Voucher Redemption
We have added a new field on the Retail Voucher Modules - "Ask for Amount". It can be ticked only on the Partial Retail Voucher module. If you try to set it to true on the other types, you're going to get an error.

![image.png](blob:https://dev.azure.com/c6c45bf6-73f0-48d6-862d-bee1b02a62a3)

If the field is set to true then when you try to pay with a voucher that has payment application method PARTIAL the system is going to ask the user to enter the amount that is going to be taken from the voucher.

![image (2).png](blob:https://dev.azure.com/821892c8-b057-4f5d-909a-f6cab86b63ed)

The suggested amount by the system is going to be the lowest amount between the Voucher's Remaining Amount and the POS Sale Amount. For example, if the Voucher Remaining Amount is 200 and the POS sale Amount is 100, then the system is going to suggest 100. If the user tries to use an amount that is bigger than the suggested amount, then the system is going to return an error that the used amount is bigger than the suggested amount.

![image (3).png](blob:https://dev.azure.com/34390d41-fdda-4c03-a5a3-c356b6368367)

Related work items: #13437