Merged PR 20505: Total Discount
New type of discount added - Total Discount.

The total discount is triggered based on the total amount of the transaction.
You can specify filters on items, item categories and Vendors for the total amount of the transaction.
that is going to trigger the discount.

Levels for the discount benefits can be set up.
The total discount can give the following benefits on different levels:
- benefit items
- discount amount
- discount percent

Related work items: #12971