Merged PR 20532: Simpler POS data refresh
This PR fixes an old backlog entry related to POS data refresh that runs at the end of every action.
Since day 1 of web POS it has been doing something roughly equivalent to the following at the end of every action click:

1. Loop all POS lines and put them into JSON.
2. Diff the new lines JSON against the last known lines JSON from last action click, each row compared against any other row twice to find both deleted and new records. (no smart sort+binary search so extra slow).
3. Send the delta between the two jsons to frontend as the "RefreshData" event.
4. Store the JSON in singleinstance/manually bound codeunit for next click to compare against.

This had several problems:
- It was slow, and it got slower the more lines was in the POS sale when clicking actions because it had to loop N lines several times, both on record and inside json.
- It was depending on state in backend and this was trying to be clever by remembering the last view. I.e. if it knew the frontend was supposed to be in payment view it would only diff the payment lines.
But this optimization just made our architecture more confusing as normally a backend should not care what is on the screen in a frontend! It should just perform the operation it is being asked to do robustly, without side effects from static state.
- It had a bug - we only refreshed data in frontend when action succeeded but all actions could do commit(); and then error later leaving behind unrefreshed data. This might have been one of the ways frontend could reach corrupt state.
- It didn't work when exposing POS actions over webservices as all single instance state is cleared between each WS call.

This PR replaces it with a simpler solution that uses table subscribers on the two POS tables to build up the delta JSON live as changes are made - and empties the buffer at the end of each action.
I know that a lot of the code between the 3 new subscribers codeunits (sale, sale line and payment line) could be shared but I am honestly so tired of premature abstractions that I kept it dead simple, 3 codeunits, 1 responsibility in each.

If we want to improve perf further we could look into fixing the TODOs I left to prevent json tranformation plus remove usage of newtonsoft as the JsonObject is not strictly optimized for writing so it has bad perf when you don't need to read.
But at least it should be constant time after this PR.

Related work items: #4335