PR 20753: New custom method for FrontEnd to get legacy POS menus
Since the idea with the new POS editor is to make the old json layouts obsolete, once a customer has switched to the new POS layouts, the old menus will no longer be loaded in on initialization. This however will be a problem because we are now implementing an old -> new layout converter which will need to read from the old menus. Rather than having these old menus load on initialization, it makes more sense to load these old menus on-demand whenever the converter is opened by the user instead.
This is what the new custom method `LegacyPOSMenus` is going to be used for.

Case 591205
Related work items: #13980