PR 20832: Reducing retention period for "NPR POS Posting Log" table from 6M to 1M
Having our retention policy for pos posting log set to 6 months does not make sense. Big data volume in the table makes looking for records in the table slower for the system, resulting in long running queries.
This is critical, because it slows down the POS posting process.
1 month as retention period for the table should be enough.

Case 591639
Related work items: #14081