Merged PR 20637: HTTP Mode
A refactoring of our frontend response logic, which is now batching everything into max 1 response per inbound request. This is both cleanup, should be easier to understand from the frontends point of view and allow us to expose the same POS objects as an HTTP API.

A new HTTP API that can receive the same payloads as our dragonglass control addin events and respond with same payloads as our dragonglass control addin procedures receive.

It depends on an currently open PR in dragonglass that allows the frontend to handle processing of an array of responses from backend. Will update this PR to contain latest dragonglass release once it merges.

It is intended to allow us to run selfservice from outside of BC - via HTTP API, so we don't have to strip away standard BC UI and maintain the logic that finds tracks it. In comparison, a static HTTP API should be much simpler to maintain long term across changing BC versions.

Related work items: #10945