PR 20844: Customer and Item rename/delete processing
Renaming a customer or an item should trigger the entity number changed elsewhere in the system, including posted transactions. Removed property `ValidateTableRelation = false` from some fields having table relation to item and customer tables. This will enable the fields to be processed by the product build-in rename routine.

Also added subscribers, which are going to kick in, when a customer or an item is deleted. System will check Np Core tables and disallow deletion, if there are unposted/unfinished entries for the customer/item.
After everything is checked, another subscriber will change the customer/item number in existing posted entries, just like this is done in standard BC.

Case 592651
Related work items: #14091