Merged PR 21046: Insert change payment line directly instead of modifying a new type onto it.
Insert change payment line directly instead of modifying a new type onto it.

There was a bug in the new data refresh when code modifies from "POS Payment" line type on POS Sale line to a sales type or vice versa.
This is something that does not really make sense to do and I don't know why the give change codeunit ever did it.

I considered putting in a hard error if anyone does it but this is not as trivial & perf free in AL as one would like:
If you use xRec in OnAfterModify you cannot trust the values.
If you use xRec in OnBeforeModify you cannot trust that a validation trigger does mess it up still and you need to re-get from the DB for all modifies of POS lines.

Instead I am leaving this gap and fixing the change code as I expect it to be very rare.
If this turns out to be wrong, we can add code in the table modify trigger, that is probably the best place.
Part of the reason I feel this very edge case gap is acceptable is because for 4-5 years there was a HUGE gap where data was not refreshed correctly if developers forgot to type POSSession.RequestRefresh() so this is nothing in comparison.

Related work items: #14251